+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 2020-07-23
 - add: _mode inherit to set the same mode to the location or code as the contact
+- fix: test if lookup exist
 
 2020-07-18
 - change binary to set